Introduced concept on invalid response from FilePicker.
Give the REST API a url which times out and you'll
get a HTTP status 200 OK back, but with a body
signalling the timeout error. It will fail with a
JSON::ParserError whenever the user of this library
tried to read response after storing the URL to find
for instance the stored file's URL.

This commit may be reverted as I'm a bit unsure of
this is the correct approach.